Add Vietnamese glyphs and ligatures to Loc_Font
Expanded Loc_Font.asset to support additional Vietnamese characters and ligatures. This includes new glyph metrics, character table entries, free glyph rects, ligature substitution records, and glyph pair adjustment records to improve font rendering for Vietnamese text.
