--- Assets/Resources_Temp/Tu_Develop.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0a7deb8dae94372419ad27a70d57be85
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
