--- Assets/Resources_Temp/Tu_Develop/Import.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8805ebdceb4aa12418f312b816578770
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
