--- Assets/Resources_Temp/Tu_Develop/Import/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 90dd174505492b14dbb7d95a2b4f8cec
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
