--- Assets/Resources_Temp/Tu_Develop/Import/Scripts/Fa/AI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 83220d75f44fa3549bd394e27bdadc1e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
