--- Assets/Resources_Temp/Tu_Develop/Import/Scripts/FaAgent.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: aaee38f0c15e9f14bb58a4d6359733ea
