--- Assets/Resources_Temp/Tu_Develop/Import/Scripts/FaInterface.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 014a853e9a43b1d4890e6331d15aad0b
