--- Assets/Scenes/Kiin_Main/Phong_scene 1/NavMesh-Test_terain 1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3aafd5ce1a94436458dbed6fc78260dc
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
