+++ Assets/Scenes/Kiin_Main/Phong_scene 1/NavMesh-Test_terain.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f92613281f95743438c685422ee5fe7a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
