--- Assets/Tu_Develop/Import/BehaviorGraph.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e2055fe14675a284c9d32490ec4185e7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
