+++ Assets/Tu_Develop/Import/Scripts/Fa.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 74cbc4e3122a84642b6f057847ba2282
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
