+++ Assets/Tu_Develop/Import/Scripts/Fa/AI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 83220d75f44fa3549bd394e27bdadc1e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
