+++ Assets/Tu_Develop/Import/Scripts/Fa/FaAgent.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 32fc6e35fe97a55469bdac3ef48138a2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
