+++ Assets/Tu_Develop/Import/Scripts/Fa/FaPrefab.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c0b84d98839ae1c42bb6fc759082341a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
